refactor retrieve functions out of searcher into retriever
