+++ lib/collection/src/collection_manager/mod.rs
@@ -3,6 +3,8 @@ pub mod holders;
 pub mod optimizers;
 pub mod segments_searcher;
 
+pub mod segments_retriever;
+
 pub mod probabilistic_search_sampling;
 mod search_result_aggregator;
